There's a looooot of is_foo, cast-to-foo calls that can/should be replaced with isinstance checks
I did them for x86SimulationState's set_register and set, but I think
if you did it for the rest of the SimLocation stuff, you'd avoid a
looot of casts.

I think you could probably go so far as to remove the is_<foo> methods
entirely, but that's totally up to you

Note that if you're suuuuper concerned with performance of the python,
you miiight want to disregard this one? I think the bool-check cast
is going to be ever so slightly faster? But if you're super concerned
with performance, this stuff should probably just be in ocaml :P