checkpoint
Moved two_way_conditionals identification from GraphInterval to a new CFG helper, since the follow node of a two way conditional is not guaranteed to be within the same interval. For example, when nesting a loop within an if, the two way conditional, the loop, and the follow will all appear in separate intervals.